# contracts/corpus-ingestion/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Norvii Corpus Ingestion API",
    "version": "1.0.0"
  },
  "servers": [{"url": "/api/v1"}],
  "paths": {
...
    "/corpora/{corpusId}/sources": {"get": {"operationId": "listSources", "parameters": [{"$ref": "#/components/parameters/CorpusId"}], "responses": {"200": {"description": "Source list", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Source"}}}}}, "404": {"$ref": "#/components/responses/Error"}}}},
    "/corpora/{corpusId}/sources/url": {"post": {"operationId": "createUrlSource", "parameters": [{"$ref": "#/components/parameters/CorpusId"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UrlSourceWrite"}}}}, "responses": {"202": {"description": "Accepted source", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Source"}}}}, "400": {"$ref": "#/components/responses/Error"}, "409": {"$ref": "#/components/responses/Error"}}}},
    "/corpora/{corpusId}/sources/pdf": {"post": {"operationId": "createPdfSource", "parameters": [{"$ref": "#/components/parameters/CorpusId"}], "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"type": "object", "required": ["title", "file"], "properties": {"title": {"type": "string"}, "file": {"type": "string", "contentMediaType": "application/pdf"}}}}}}, "responses": {"202": {"description": "Accepted source", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Source"}}}}, "413": {"$ref": "#/components/responses/Error"}}}},
    "/corpora/{corpusId}/sources/{sourceId}": {"get": {"operationId": "getSource", "parameters": [{"$ref": "#/components/parameters/CorpusId"}, {"$ref": "#/components/parameters/SourceId"}], "responses": {"200": {"description": "Source", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Source"}}}}, "404": {"$ref": "#/components/responses/Error"}}}},
    "/corpora/{corpusId}/sources/{sourceId}/retry": {"post": {"operationId": "retrySource", "parameters": [{"$ref": "#/components/parameters/CorpusId"}, {"$ref": "#/components/parameters/SourceId"}], "requestBody": {"$ref": "#/components/requestBodies/VersionCommand"}, "responses": {"202": {"description": "Retry accepted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Source"}}}}, "409": {"$ref": "#/components/responses/Error"}}}},
    "/corpora/{corpusId}/sources/{sourceId}/reprocess": {"post": {"operationId": "reprocessSource", "parameters": [{"$ref": "#/components/parameters/CorpusId"}, {"$ref": "#/components/parameters/SourceId"}], "requestBody": {"$ref": "#/components/requestBodies/VersionCommand"}, "responses": {"202": {"description": "Reprocess accepted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Source"}}}}, "409": {"$ref": "#/components/responses/Error"}}}},
...
    "/corpora/{corpusId}/sources/{sourceId}/origin/url": {"get": {"operationId": "getUrlOrigin", "parameters": [{"$ref": "#/components/parameters/CorpusId"}, {"$ref": "#/components/parameters/SourceId"}], "responses": {"200": {"description": "Captured URL origin metadata", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SourceOrigin"}}}}, "404": {"$ref": "#/components/responses/Error"}}}},
    "/corpora/{corpusId}/sources/{sourceId}/origin/pdf": {"get": {"operationId": "downloadPdfOrigin", "parameters": [{"$ref": "#/components/parameters/CorpusId"}, {"$ref": "#/components/parameters/SourceId"}], "responses": {"200": {"description": "Preserved PDF origin", "content": {"application/pdf": {"schema": {"type": "string", "contentMediaType": "application/pdf"}}}}, "404": {"$ref": "#/components/responses/Error"}}}}
  },
  "components": {
    "parameters": {
      "CorpusId": {"name": "corpusId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
      "SourceId": {"name": "sourceId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
    },
    "requestBodies": {"VersionCommand": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["version"], "properties": {"version": {"type": "integer", "minimum": 1}}}}}}},
    "responses": {"Error": {"description": "Safe error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}},
    "schemas": {
...
      "UrlSourceWrite": {"type": "object", "additionalProperties": false, "required": ["title", "url"], "properties": {"title": {"type": "string", "minLength": 1}, "url": {"type": "string", "format": "uri", "pattern": "^https://"}}},
      "Source": {"type": "object", "required": ["id", "corpusId", "title", "kind", "processingStatus", "version", "createdAt", "updatedAt", "origin", "latestAttempt", "attempts"], "properties": {"id": {"type": "string", "format": "uuid"}, "corpusId": {"type": "string", "format": "uuid"}, "title": {"type": "string"}, "kind": {"type": "string", "enum": ["url", "pdf"]}, "processingStatus": {"type": "string", "enum": ["pending", "processing", "ready", "failed"]}, "failureCategory": {"type": ["string", "null"]}, "latestReadyDocumentId": {"type": ["string", "null"], "format": "uuid"}, "version": {"type": "integer", "minimum": 1}, "createdAt": {"type": "string", "format": "date-time"}, "updatedAt": {"type": "string", "format": "date-time"}, "origin": {"$ref": "#/components/schemas/SourceOrigin"}, "latestAttempt": {"oneOf": [{"$ref": "#/components/schemas/ProcessingAttempt"}, {"type": "null"}]}, "attempts": {"type": "array", "items": {"$ref": "#/components/schemas/ProcessingAttempt"}}}},
      "SourceOrigin": {"type": "object", "required": ["kind"], "properties": {"kind": {"type": "string", "enum": ["url", "pdf"]}, "submittedUrl": {"type": "string", "format": "uri"}, "normalizedUrl": {"type": "string", "format": "uri"}, "originalFilename": {"type": "string"}, "mediaType": {"type": "string"}, "byteSize": {"type": "integer", "minimum": 0}, "sha256": {"type": "string", "pattern": "^[0-9a-f]{64}$"}, "finalUrl": {"type": "string", "format": "uri"}, "capturedAt": {"type": "string", "format": "date-time"}, "extractedContentSha256": {"type": "string", "pattern": "^[0-9a-f]{64}$"}}},
      "ProcessingAttempt": {"type": "object", "required": ["number", "pipelineVersion", "status", "startedAt"], "properties": {"number": {"type": "integer", "minimum": 1}, "pipelineVersion": {"type": "string"}, "status": {"type": "string", "enum": ["processing", "succeeded", "failed"]}, "startedAt": {"type": "string", "format": "date-time"}, "finishedAt": {"type": ["string", "null"], "format": "date-time"}, "failureCategory": {"type": ["string", "null"]}, "acquiredByteCount": {"type": ["integer", "null"], "minimum": 0}, "normalizedCharacterCount": {"type": ["integer", "null"], "minimum": 0}, "unitCount": {"type": ["integer", "null"], "minimum": 0}, "durationMilliseconds": {"type": ["integer", "null"], "minimum": 0}}},
...
      "ErrorEnvelope": {"type": "object", "required": ["error"], "properties": {"error": {"type": "object", "required": ["code", "message", "requestId"], "properties": {"code": {"type": "string", "enum": ["invalid_input", "payload_too_large", "unsafe_url", "unsupported_content", "duplicate_source", "stale_state", "not_found", "unavailable", "acquisition_failed", "extraction_failed", "publication_failed", "internal_error"]}, "message": {"type": "string"}, "fields": {"type": "object", "additionalProperties": {"type": "string"}}, "requestId": {"type": "string", "format": "uuid"}}}}
    }
  }
}
}